+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add bottom title to `Block` (`Block::title_bottom`)
 - Add support for multiple titles to `Block`
 - Add `row_span` and `col_span` support to `Grid`
+- Add `label` to `ProgressBar`
 
 ### Changes: